cover: refact -> add edit.js. Change cpms names
